Moved dropHead method in HPUtils to EntityDeathListener
couldn't shorten to EDL for obvious reasons
